requirements.txt
click>=8.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
genanki>=0.13.0
numpy>=1.21.0
scipy>=1.10.0
opencv-python>=4.0.0
Pillow>=9.0.0
pillow-heif>=0.10.0
easyocr>=1.7.0
google-generativeai>=0.3.0
sudachipy>=0.6.0
sudachidict_core>=20210802
streamlit>=1.30.0
